Revert shadow picker for ShadowResources, for now.
Lots of people extend ShadowResources for their own nefarious purposes.
We'll address that later.
