+++ robolectric/src/test/java/org/robolectric/shadows/ShadowAssetManagerTest.java
@@ -5,7 +5,6 @@ import static org.junit.Assert.fail;
 import static org.junit.Assume.assumeTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static org.robolectric.Shadows.shadowOf;
 import static org.robolectric.shadows.ShadowAssetManager.legacyShadowOf;
 import static org.robolectric.shadows.ShadowAssetManager.useLegacy;
 
@@ -28,7 +27,7 @@ import org.robolectric.RobolectricTestRunner;
 import org.robolectric.RuntimeEnvironment;
 import org.robolectric.annotation.Config;
 import org.robolectric.shadow.api.Shadow;
-import org.robolectric.shadows.ShadowLegacyResources.ShadowLegacyTheme;
+import org.robolectric.shadows.ShadowResources.ShadowLegacyTheme;
 
 @RunWith(RobolectricTestRunner.class)
 public class ShadowAssetManagerTest {
